自定义用例模块和查询判断结果

+++ foj-springboot/foj-commons/src/main/java/icu/chzqz/foj/entity/Question.java
@@ -21,7 +21,7 @@ public class Question {
     private Long maxTime;
     private Long maxMemory;
     private Long maxProc;
-    private Long uid;
+    private Integer uid;
     private Integer status;
     private Integer level;
 }
